feat: Refactor Cart and Order Adapters to use BigDecimal for price calculations
feat: Update ApiService to include MultipartBody for image uploads

feat: Modify GenerativeModel to remove API key dependency

feat: Enhance CartItem model to include an ID field

feat: Update Product model to support multiple categories and availability status

fix: Improve AddEditCategoryActivity to handle image URI more effectively

fix: Adjust AddEditProductActivity to use FileUtil for image path retrieval

feat: Update AdminProductListActivity to accommodate new ProductAdapter constructor

fix: Refactor BaseActivity to streamline CartRepository initialization

feat: Implement CategoryListActivity to load categories from the repository

feat: Enhance CategoryProductListActivity to utilize new ProductAdapter

fix: Update ChatActivity to use ProductAdapter for product interactions

feat: Modify HomeActivity to handle product addition to cart

fix: Update LoginActivity to store user's full name correctly

fix: Adjust ProductDetailActivity to initialize CartRepository properly

fix: Refactor RegisterActivity to simplify user creation

refactor: Update Category and User models for better structure

feat: Implement FileUtil for URI to file path conversion

style: Update product_item layout to include an "Add to cart" button

chore: Remove unused API key from build.gradle.kts

chore: Add environment variable template for sensitive keys

+++ fe/app/src/main/java/com/nstut/fast_food_shop/util/Utils.java
@@ -1,10 +1,11 @@
 package com.nstut.fast_food_shop.util;
 
+import java.math.BigDecimal;
 import java.text.NumberFormat;
 import java.util.Locale;
 
 public class Utils {
-    public static String formatCurrency(double amount) {
+    public static String formatCurrency(BigDecimal amount) {
         NumberFormat formatter = NumberFormat.getCurrencyInstance(new Locale("en", "US"));
         return formatter.format(amount);
     }
